feat(event-queues): [CHK-4193] add service separation of concerns
