+++ src/test/kotlin/it/pagopa/helpdeskcommands/services/CommandsServiceTest.kt
@@ -1,27 +1,6 @@
 package it.pagopa.helpdeskcommands.services
 
-import com.azure.core.http.rest.Response as AzureResponse
-import com.azure.storage.queue.models.SendMessageResult
 import com.fasterxml.jackson.databind.ObjectMapper
-import io.opentelemetry.api.OpenTelemetry
-import io.opentelemetry.api.trace.Span
-import io.opentelemetry.api.trace.SpanBuilder
-import io.opentelemetry.api.trace.SpanContext
-import io.opentelemetry.api.trace.SpanKind
-import io.opentelemetry.api.trace.Tracer
-import io.opentelemetry.context.Context
-import io.opentelemetry.context.propagation.ContextPropagators
-import io.opentelemetry.context.propagation.TextMapPropagator
-import io.opentelemetry.context.propagation.TextMapSetter
-import it.pagopa.ecommerce.commons.client.QueueAsyncClient
-import it.pagopa.ecommerce.commons.documents.v2.TransactionRefundRequestedData
-import it.pagopa.ecommerce.commons.documents.v2.TransactionRefundRequestedEvent
-import it.pagopa.ecommerce.commons.documents.v2.TransactionUserReceiptData
-import it.pagopa.ecommerce.commons.documents.v2.TransactionUserReceiptRequestedEvent
-import it.pagopa.ecommerce.commons.generated.server.model.TransactionStatusDto
-import it.pagopa.ecommerce.commons.queues.QueueEvent
-import it.pagopa.ecommerce.commons.queues.TracingInfo
-import it.pagopa.ecommerce.commons.queues.TracingUtils
 import it.pagopa.generated.ecommerce.redirect.v1.dto.RefundRequestDto as RedirectRefundRequestDto
 import it.pagopa.generated.ecommerce.redirect.v1.dto.RefundResponseDto as RedirectRefundResponseDto
 import it.pagopa.generated.helpdeskcommands.model.RefundOutcomeDto
@@ -30,43 +9,23 @@ import it.pagopa.generated.npg.api.PaymentServicesApi
 import it.pagopa.helpdeskcommands.client.NodeForwarderClient
 import it.pagopa.helpdeskcommands.client.NpgClient
 import it.pagopa.helpdeskcommands.config.WebClientConfig
-import it.pagopa.helpdeskcommands.exceptions.NodeForwarderClientException
-import it.pagopa.helpdeskcommands.exceptions.NpgApiKeyConfigurationException
-import it.pagopa.helpdeskcommands.exceptions.NpgClientException
-import it.pagopa.helpdeskcommands.exceptions.RedirectConfigurationException
-import it.pagopa.helpdeskcommands.utils.NpgApiKeyConfiguration
-import it.pagopa.helpdeskcommands.utils.NpgPspApiKeysConfig
-import it.pagopa.helpdeskcommands.utils.PaymentMethod
-import it.pagopa.helpdeskcommands.utils.RedirectKeysConfiguration
-import it.pagopa.helpdeskcommands.utils.TransactionId
+import it.pagopa.helpdeskcommands.exceptions.*
+import it.pagopa.helpdeskcommands.utils.*
 import java.math.BigDecimal
 import java.net.URI
-import java.time.Duration
 import java.util.*
 import java.util.stream.Stream
 import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
 import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Assertions.assertNotNull
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.BeforeAll
-import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
-import org.mockito.ArgumentMatchers.anyString
-import org.mockito.Mockito
-import org.mockito.kotlin.any
-import org.mockito.kotlin.argumentCaptor
-import org.mockito.kotlin.eq
-import org.mockito.kotlin.given
-import org.mockito.kotlin.mock
-import org.mockito.kotlin.reset
-import org.mockito.kotlin.spy
-import org.mockito.kotlin.times
-import org.mockito.kotlin.verify
+import org.mockito.kotlin.*
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.http.HttpStatus
 import org.springframework.test.context.TestPropertySource
@@ -115,74 +74,13 @@ class CommandsServiceTest {
             redirectBeAoiCallUriSet
         )
 
-    private val refundQueueClient: QueueAsyncClient = mock()
-    private val notificationQueueClient: QueueAsyncClient = mock()
-    private val transientQueueTTLSeconds: Long = 2700 // 45 minutes
-    private lateinit var commandsService: CommandsService
-    private lateinit var spiedTracingUtils: TracingUtils
-
-    @BeforeEach
-    fun setupWithRealTracingUtils() {
-        reset(npgClient, nodeForwarderRedirectApiClient, refundQueueClient, notificationQueueClient)
-
-        val mockOpenTelemetry: OpenTelemetry = mock()
-        val mockTracer: Tracer = mock()
-        val mockSpanBuilder: SpanBuilder = mock()
-        val mockSpan: Span = mock()
-        val mockContextPropagators: ContextPropagators = mock()
-        val mockTextMapPropagator: TextMapPropagator = mock()
-
-        given(mockOpenTelemetry.getTracer(anyString())).willReturn(mockTracer)
-        given(mockOpenTelemetry.propagators).willReturn(mockContextPropagators)
-        given(mockContextPropagators.textMapPropagator).willReturn(mockTextMapPropagator)
-
-        Mockito.doAnswer { invocation ->
-                val carrier = invocation.getArgument<HashMap<String, String>>(1)
-                @Suppress("UNCHECKED_CAST")
-                val setter = invocation.getArgument<TextMapSetter<HashMap<String, String>>>(2)
-
-                setter.set(carrier, TracingUtils.TRACEPARENT, "00-dummytraceid-dummyspanid-01")
-                setter.set(carrier, TracingUtils.TRACESTATE, "statekey=statevalue")
-                setter.set(carrier, TracingUtils.BAGGAGE, "baggagekey=baggagevalue")
-                null
-            }
-            .`when`(mockTextMapPropagator)
-            .inject(
-                any<Context>(),
-                any<HashMap<String, String>>(),
-                any<TextMapSetter<HashMap<String, String>>>()
-            )
-
-        given(mockTextMapPropagator.fields())
-            .willReturn(
-                listOf(TracingUtils.TRACEPARENT, TracingUtils.TRACESTATE, TracingUtils.BAGGAGE)
-            )
-
-        given(mockTracer.spanBuilder(anyString())).willReturn(mockSpanBuilder)
-
-        given(mockSpanBuilder.setSpanKind(any<SpanKind>())).willReturn(mockSpanBuilder)
-        given(mockSpanBuilder.setParent(any<Context>())).willReturn(mockSpanBuilder)
-        given(mockSpanBuilder.setNoParent()).willReturn(mockSpanBuilder)
-        given(mockSpanBuilder.startSpan()).willReturn(mockSpan)
-
-        Mockito.`when`(mockSpan.spanContext).thenReturn(SpanContext.getInvalid())
-        given(mockSpan.isRecording).willReturn(true)
-
-        val actualTracingUtilsInstance = TracingUtils(mockOpenTelemetry, mockTracer)
-        this.spiedTracingUtils = spy(actualTracingUtilsInstance)
-
-        commandsService =
-            CommandsService(
-                npgClient = npgClient,
-                npgApiKeyConfiguration = npgApiKeyConfiguration,
-                redirectKeysConfiguration = redirectKeysConfiguration,
-                nodeForwarderClient = nodeForwarderRedirectApiClient,
-                refundQueueClient = Mono.just(refundQueueClient),
-                notificationQueueClient = Mono.just(notificationQueueClient),
-                transientQueueTTLSeconds = transientQueueTTLSeconds,
-                tracingUtils = this.spiedTracingUtils
-            )
-    }
+    private val commandsService: CommandsService =
+        CommandsService(
+            npgClient = npgClient,
+            npgApiKeyConfiguration = npgApiKeyConfiguration,
+            redirectKeysConfiguration = redirectKeysConfiguration,
+            nodeForwarderClient = nodeForwarderRedirectApiClient
+        )
 
     companion object {
         lateinit var mockWebServer: MockWebServer
@@ -384,11 +282,7 @@ class CommandsServiceTest {
                 npgClient = npgClient,
                 npgApiKeyConfiguration = npgApiKeyConfiguration,
                 redirectKeysConfiguration = redirectKeysConfiguration,
-                nodeForwarderClient = nodeForwarderRedirectApiClient,
-                refundQueueClient = Mono.just(mock()),
-                notificationQueueClient = Mono.just(mock()),
-                transientQueueTTLSeconds = 2700,
-                tracingUtils = mock()
+                nodeForwarderClient = nodeForwarderRedirectApiClient
             )
         val operationId = "operationID"
         val transactionId = TransactionId(TRANSACTION_ID_STRING)
@@ -438,11 +332,7 @@ class CommandsServiceTest {
                 npgClient = npgClient,
                 npgApiKeyConfiguration = npgApiKeyConfiguration,
                 redirectKeysConfiguration = redirectKeysConfiguration,
-                nodeForwarderClient = nodeForwarderRedirectApiClient,
-                refundQueueClient = Mono.just(mock()),
-                notificationQueueClient = Mono.just(mock()),
-                transientQueueTTLSeconds = 2700,
-                tracingUtils = mock()
+                nodeForwarderClient = nodeForwarderRedirectApiClient
             )
         val operationId = "operationID"
         val transactionId = TransactionId(TRANSACTION_ID_STRING)
@@ -492,11 +382,7 @@ class CommandsServiceTest {
                 npgClient = npgClient,
                 npgApiKeyConfiguration = npgApiKeyConfiguration,
                 redirectKeysConfiguration = redirectKeysConfiguration,
-                nodeForwarderClient = nodeForwarderRedirectApiClient,
-                refundQueueClient = Mono.just(mock()),
-                notificationQueueClient = Mono.just(mock()),
-                transientQueueTTLSeconds = 2700,
-                tracingUtils = mock()
+                nodeForwarderClient = nodeForwarderRedirectApiClient
             )
         val operationId = "operationID"
         val transactionId = TransactionId(TRANSACTION_ID_STRING)
@@ -651,157 +537,5 @@ class CommandsServiceTest {
                 transactionId,
                 RedirectRefundResponseDto::class.java
             )
-    }
-
-    @Test
-    fun `should send refund requested event successfully`() {
-        val transactionId = TRANSACTION_ID_STRING
-        val refundEvent =
-            TransactionRefundRequestedEvent(
-                transactionId,
-                TransactionRefundRequestedData(
-                    null,
-                    TransactionStatusDto.CLOSED,
-                    TransactionRefundRequestedData.RefundTrigger.MANUAL
-                )
-            )
-
-        val mockAzureResponse: AzureResponse<SendMessageResult> = mock()
-        val queueEventCaptor = argumentCaptor<QueueEvent<TransactionRefundRequestedEvent>>()
-
-        given(
-                refundQueueClient.sendMessageWithResponse(
-                    any<QueueEvent<TransactionRefundRequestedEvent>>(),
-                    eq(Duration.ZERO),
-                    eq(Duration.ofSeconds(transientQueueTTLSeconds))
-                )
-            )
-            .willReturn(Mono.just(mockAzureResponse))
-
-        StepVerifier.create(commandsService.sendRefundRequestedEvent(refundEvent)).verifyComplete()
-
-        verify(spiedTracingUtils)
-            .traceMono(
-                eq(CommandsService::class.java.simpleName),
-                any<java.util.function.Function<TracingInfo, Mono<Void>>>()
-            )
-
-        verify(refundQueueClient)
-            .sendMessageWithResponse(
-                queueEventCaptor.capture(),
-                eq(Duration.ZERO),
-                eq(Duration.ofSeconds(transientQueueTTLSeconds))
-            )
-
-        val capturedQueueEvent = queueEventCaptor.firstValue
-        assertEquals(refundEvent, capturedQueueEvent.event)
-        assertNotNull(
-            capturedQueueEvent.tracingInfo,
-            "TracingInfo should be populated in QueueEvent"
-        )
-    }
-
-    @Test
-    fun `should handle queue error when sending refund event`() {
-        val transactionId = TRANSACTION_ID_STRING
-        val event =
-            TransactionRefundRequestedEvent(
-                transactionId,
-                TransactionRefundRequestedData(
-                    null,
-                    TransactionStatusDto.CLOSED,
-                    TransactionRefundRequestedData.RefundTrigger.MANUAL
-                )
-            )
-
-        given(
-                refundQueueClient.sendMessageWithResponse(
-                    any<QueueEvent<TransactionRefundRequestedEvent>>(),
-                    any<Duration>(),
-                    any<Duration>()
-                )
-            )
-            .willReturn(Mono.error(RuntimeException("Queue error")))
-
-        StepVerifier.create(commandsService.sendRefundRequestedEvent(event))
-            .expectError(RuntimeException::class.java)
-            .verify()
-    }
-
-    @Test
-    fun `should send notification requested event successfully`() {
-        val transactionId = TRANSACTION_ID_STRING
-
-        val userReceiptData =
-            TransactionUserReceiptData(
-                TransactionUserReceiptData.Outcome.OK,
-                "IT",
-                "2023-10-26T10:00:00",
-                TransactionUserReceiptData.NotificationTrigger.MANUAL
-            )
-
-        val notificationEvent = TransactionUserReceiptRequestedEvent(transactionId, userReceiptData)
-
-        val mockAzureResponse: AzureResponse<SendMessageResult> = mock()
-        val queueEventCaptor = argumentCaptor<QueueEvent<TransactionUserReceiptRequestedEvent>>()
-
-        given(
-                notificationQueueClient.sendMessageWithResponse(
-                    any<QueueEvent<TransactionUserReceiptRequestedEvent>>(),
-                    eq(Duration.ZERO),
-                    eq(Duration.ofSeconds(transientQueueTTLSeconds))
-                )
-            )
-            .willReturn(Mono.just(mockAzureResponse))
-
-        StepVerifier.create(commandsService.sendNotificationRequestedEvent(notificationEvent))
-            .verifyComplete()
-
-        verify(spiedTracingUtils)
-            .traceMono(
-                eq(CommandsService::class.java.simpleName),
-                any<java.util.function.Function<TracingInfo, Mono<Void>>>()
-            )
-
-        verify(notificationQueueClient)
-            .sendMessageWithResponse(
-                queueEventCaptor.capture(),
-                eq(Duration.ZERO),
-                eq(Duration.ofSeconds(transientQueueTTLSeconds))
-            )
-
-        val capturedQueueEvent = queueEventCaptor.firstValue
-        assertEquals(notificationEvent, capturedQueueEvent.event)
-        assertNotNull(
-            capturedQueueEvent.tracingInfo,
-            "TracingInfo should be populated in QueueEvent"
-        )
-    }
-
-    @Test
-    fun `should handle queue error when sending notification event`() {
-        val transactionId = TRANSACTION_ID_STRING
-
-        val userReceiptDataForError =
-            TransactionUserReceiptData(
-                TransactionUserReceiptData.Outcome.OK,
-                "IT",
-                "2023-10-26T10:00:00",
-                TransactionUserReceiptData.NotificationTrigger.MANUAL
-            )
-        val event = TransactionUserReceiptRequestedEvent(transactionId, userReceiptDataForError)
-
-        given(
-                notificationQueueClient.sendMessageWithResponse(
-                    any<QueueEvent<TransactionUserReceiptRequestedEvent>>(),
-                    any<Duration>(),
-                    any<Duration>()
-                )
-            )
-            .willReturn(Mono.error(RuntimeException("Queue error")))
-
-        StepVerifier.create(commandsService.sendNotificationRequestedEvent(event))
-            .expectError(RuntimeException::class.java)
-            .verify()
     }
 }
